+++ Cosette/Engine/Ai/Ordering/MoveOrderingConstants.cs
@@ -2,11 +2,11 @@
 {
     public static class MoveOrderingConstants
     {
-        public const int HashMove = 10000;
-        public const int Promotion = 5000;
-        public const int Capture = 100;
-        public const int KillerMove = 80;
-        public const int PawnNearPromotion = 50;
+        public const short HashMove = 10000;
+        public const short Promotion = 5000;
+        public const short Capture = 100;
+        public const short KillerMove = 80;
+        public const short PawnNearPromotion = 50;
 
         public const int KillerSlots = 3;
     }
